BACK-TalentoTech/requirements.txt
FastAPI
pandas
matplotlib
uvicorn
pillow